More fixes and cleanup

+++ Repository/Interface/ICommentRepository.cs
@@ -12,6 +12,6 @@ namespace Clara.Repository.Interface
         bool HasUserComment(string userId, Guid serviceId);
         IEnumerable<Comment> GetComments(Guid serviceId);
         IEnumerable<Comment> GetUserComments(string userId);
-        IEnumerable<double> GetTotalServiceRating(Guid serviceId);
+        IEnumerable<int> GetTotalServiceRating(Guid serviceId);
     }
 }
